requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=mkl
abseil-cpp=20211102.0=h0e60522_0
absl-py=1.3.0=pyhd8ed1ab_0
aiohttp=3.8.1=py310he2412df_1
aiosignal=1.3.1=pyhd8ed1ab_0
anyio=3.5.0=py310haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py310h2bbff1b_0
asttokens=2.0.5=pyhd3eb1b0_0
astunparse=1.6.3=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=22.1.0=pyh71513ae_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.11.1=py310haa95532_0
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.5=pyhd8ed1ab_0
brotlipy=0.7.0=py310he2412df_1004
bzip2=1.0.8=he774522_0
ca-certificates=2022.12.7=h5b45459_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.2.0=pyhd8ed1ab_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310hcbf9ad4_0
charset-normalizer=2.1.1=pyhd8ed1ab_0
click=8.1.3=win_pyhd8ed1ab_2
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.0.6=pypi_0
cryptography=37.0.4=py310ha857299_0
cycler=0.11.0=pypi_0
debugpy=1.5.1=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
eigen=3.4.0=h2d74725_0
entrypoints=0.4=py310haa95532_0
executing=0.8.3=pyhd3eb1b0_0
flask=2.2.2=pypi_0
flask-cors=3.0.10=pypi_0
flatbuffers=2.0.0=h6c2663c_0
fonttools=4.38.0=pypi_0
frozenlist=1.3.3=py310h2bbff1b_0
gast=0.5.3=pyhd8ed1ab_0
giflib=5.2.1=h8d14728_2
glib=2.69.1=h5dc1a3c_2
google-auth=2.15.0=pyh1a96a4e_0
google-auth-oauthlib=0.4.6=pyhd8ed1ab_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.42.0=py310hc60d5dd_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h5py=3.7.0=nompi_py310h00cbb18_100
hdf5=1.12.1=nompi_h2a0e4a3_100
icu=58.2=ha925a31_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=5.2.0=pyha770c72_0
imutils=0.5.4=pypi_0
intel-openmp=2023.0.0=h57928b3_25922
ipykernel=6.15.2=py310haa95532_0
ipython=8.7.0=py310haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
itsdangerous=2.1.2=pypi_0
jedi=0.18.1=py310haa95532_1
jinja2=3.1.2=py310haa95532_0
jpeg=9e=h8ffe710_2
jsonschema=4.16.0=py310haa95532_0
jupyter_client=7.4.7=py310haa95532_0
jupyter_core=4.11.2=py310haa95532_0
jupyter_server=1.18.1=py310haa95532_0
jupyterlab_pygments=0.1.2=py_0
keras=2.9.0=py310haa95532_0
keras-preprocessing=1.1.2=pyhd8ed1ab_0
kiwisolver=1.4.4=pypi_0
libblas=3.9.0=8_mkl
libcblas=3.9.0=8_mkl
libclang=12.0.0=default_h627e005_2
libcurl=7.86.0=h86230a5_0
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
liblapack=3.9.0=8_mkl
libogg=1.3.4=h8ffe710_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.1=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=h680486a_2
libtiff=4.0.9=h36446d0_1002
libvorbis=1.3.7=h0e60522_0
libwebp=1.2.4=h8ffe710_0
libwebp-base=1.2.4=h8ffe710_0
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lxml=4.9.1=py310h1985fb9_0
lz4-c=1.9.3=h8ffe710_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.4.1=pyhd8ed1ab_0
markupsafe=2.1.1=py310he2412df_1
matplotlib=3.6.2=pypi_0
matplotlib-inline=0.1.6=py310haa95532_0
mistune=0.8.4=py310h2bbff1b_1000
mkl=2020.4=hb70f87d_311
msys2-conda-epoch=20160418=1
multidict=6.0.2=py310he2412df_1
nbclassic=0.4.8=py310haa95532_0
nbclient=0.5.13=py310haa95532_0
nbconvert=6.5.4=py310haa95532_0
nbformat=5.7.0=py310haa95532_0
nest-asyncio=1.5.5=py310haa95532_0
notebook=6.5.2=py310haa95532_0
notebook-shim=0.2.2=py310haa95532_0
numpy=1.22.4=py310hed7ac4c_0
oauthlib=3.2.2=pyhd8ed1ab_0
opencv=4.6.0=py310ha7641e4_2
openssl=1.1.1s=h2bbff1b_0
opt_einsum=3.3.0=pyhd8ed1ab_1
packaging=22.0=pyhd8ed1ab_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h0e60522_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.3.0=pypi_0
pip=22.3.1=py310haa95532_0
prometheus_client=0.14.1=py310haa95532_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.20.1=py310hd77b12b_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.11.2=pyhd3eb1b0_0
pyjwt=2.6.0=pyhd8ed1ab_0
pyopenssl=22.0.0=pyhd8ed1ab_1
pyparsing=3.0.9=pypi_0
pyrsistent=0.18.0=py310h2bbff1b_0
pysocks=1.7.1=pyh0701188_6
python=3.10.8=h966fe2a_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310haa95532_0
python-flatbuffers=22.12.6=pyhd8ed1ab_0
python_abi=3.10=2_cp310
pyu2f=0.1.5=pyhd8ed1ab_0
pywin32=305=py310h2bbff1b_0
pywinpty=2.0.2=py310h5da7b33_0
pyyaml=6.0=py310he2412df_4
pyzmq=23.2.0=py310hd77b12b_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_4
qtwebkit=5.212=h3ad3cdb_4
requests=2.28.1=pyhd8ed1ab_1
requests-oauthlib=1.3.1=pyhd8ed1ab_0
rsa=4.9=pyhd8ed1ab_0
scipy=1.7.3=py310h33db832_0
send2trash=1.8.0=pyhd3eb1b0_1
setuptools=65.5.0=py310haa95532_0
six=1.16.0=pyh6c4a22f_0
snappy=1.1.9=h82413e6_1
sniffio=1.2.0=py310haa95532_1
soupsieve=2.3.2.post1=py310haa95532_0
sqlite=3.40.0=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tensorboard=2.9.0=py310haa95532_0
tensorboard-data-server=0.6.1=py310h5588dad_4
tensorboard-plugin-wit=1.8.1=pyhd8ed1ab_0
tensorflow=2.9.1=mkl_py310h626feff_1
tensorflow-base=2.9.1=mkl_py310h6a7f48e_1
tensorflow-estimator=2.9.0=py310haa95532_0
termcolor=2.1.1=pyhd8ed1ab_0
terminado=0.13.1=py310haa95532_0
tinycss2=1.2.1=py310haa95532_0
tk=8.6.12=h2bbff1b_0
tornado=6.2=py310h2bbff1b_0
traitlets=5.7.1=py310haa95532_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.13=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py310haa95532_1
websocket-client=0.58.0=py310haa95532_4
werkzeug=2.2.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd3eb1b0_0
win_inet_pton=1.1.0=pyhd8ed1ab_6
wincertstore=0.2=py310haa95532_2
winpty=0.4.3=4
wrapt=1.14.1=py310he2412df_0
xz=5.2.8=h8cc25b3_0
yaml=0.2.5=h8ffe710_2
yarl=1.7.2=py310he2412df_2
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=pyhd8ed1ab_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
